updated the location of the config file
